added multimodule maven project and profiling
