+++ api-gateway/src/main/java/com/service/apigateway/ApiGatewayApplication.java
@@ -1,5 +1,6 @@
-package com.howtodoinjava.example.apigateway;
+package com.service.apigateway;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -11,10 +12,11 @@ import org.springframework.cloud.netflix.zuul.EnableZuulServer;
 @EnableEurekaClient
 @EnableDiscoveryClient
 @EnableZuulProxy
-@EnableZuulServer
+@Slf4j
 public class ApiGatewayApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(ApiGatewayApplication.class, args);
+        log.debug("Api Gateway Application started..!!");
     }
 }
