+++ api-gateway/src/main/resources/application-prod.yml
@@ -1,0 +1,33 @@
+spring:
+  application:
+    name: api-gateway
+
+server:
+  port: 8010
+
+eureka:
+  instance:
+    hostname: api-gateway
+  client:
+    registerWithEureka: true
+    fetchRegistry: true
+    serviceUrl:
+      defaultZone: ${EUREKA_URI:http://eureka-server.default.svc.cluster.local:8761/eureka}
+    healthcheck:
+      enabled: true
+    lease:
+      duration: 5
+
+zuul:
+  routes:
+    employee-service:
+      path: /employee-service/**
+      serviceId: employee-service
+
+management:
+  security:
+    enabled: false
+
+logging:
+  level:
+    com.service.apigateway: DEBUG
