+++ api-gateway/src/main/resources/application.yml
@@ -1,31 +1,3 @@
-server:
-  port: 8010    
-
-eureka:
-  instance:
-    hostname: api-gateway
-  client:
-    registerWithEureka : true
-    fetchRegistry : true
-    serviceUrl:
-      defaultZone: ${EUREKA_URI:http://eureka-server.default.svc.cluster.local:8761/eureka}
-    healthcheck:
-      enabled: true
-    lease:
-      duration: 5
-
 spring:
-  application:
-    name: api-gateway
-zuul:
-  routes:
-    employee-service: 
-          path: /employee-service/**
-          serviceId: employee-service
-management:
-  security:
-    enabled: false
-
-logging:
-  level:
-    com.self.sprintboot.learning.apigateway: DEBUG
+  profiles:
+  active: "dev"
