--- employee-service/.mvn/wrapper/maven-wrapper.properties
@@ -1,1 +1,0 @@
-distributionUrl=https://repo1.maven.org/maven2/org/apache/maven/apache-maven/3.5.2/apache-maven-3.5.2-bin.zip
