+++ employee-service/Dockerfile
@@ -1,5 +1,5 @@
 FROM openjdk:8-jdk-alpine
-COPY target/com.self.sprintboot.learning.employee-0.0.1-SNAPSHOT.jar /var/app/application.jar
+COPY target/employee-service-0.0.1-SNAPSHOT.jar /var/app/application.jar
 ENTRYPOINT ["/usr/bin/java"]
 CMD ["-jar", "/var/app/application.jar"]
 VOLUME /var/app
