+++ employee-service/src/main/java/com/service/employee/EmployeeServiceApplication.java
@@ -1,14 +1,17 @@
-package com.howtodoinjava.example.employee;
+package com.service.employee;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @SpringBootApplication
 @EnableEurekaClient
+@Slf4j
 public class EmployeeServiceApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(EmployeeServiceApplication.class, args);
+		log.debug("Api Gateway Application started..!!");
 	}
 }
