+++ employee-service/src/main/java/com/service/employee/beans/Employee.java
@@ -1,0 +1,13 @@
+package com.service.employee.beans;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Employee {
+    private int id;
+    private String name;
+}
