+++ employee-service/src/main/resources/application-dev.yml
@@ -1,0 +1,28 @@
+spring:
+  application:
+    name: employee-service
+
+server:
+  port: 8011
+
+eureka:
+    instance:
+      hostname: employee-service
+    client:
+      serviceUrl:
+        defaultZone: http://localhost:8761/eureka
+      healthcheck:
+        enabled: true
+      lease:
+        duration: 5
+
+management:
+  security:
+    enabled: false
+
+logging:
+  level:
+    com.service.employee: DEBUG
+
+
+
