+++ eureka-server/Dockerfile
@@ -1,5 +1,5 @@
 FROM openjdk:8-jdk-alpine
-COPY target/com-self-sprintboot-learning-eureka-0.0.1-SNAPSHOT.jar /var/app/application.jar
+COPY target/eureka-server-0.0.1-SNAPSHOT.jar /var/app/application.jar
 ENTRYPOINT ["/usr/bin/java"]
 CMD ["-jar", "/var/app/application.jar"]
 VOLUME /var/app
