--- eureka-server/src/main/java/com/howtodoinjava/example/eureka/EurekaServerApplication.java
@@ -1,18 +1,0 @@
-package com.howtodoinjava.example.eureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-import de.codecentric.boot.admin.config.EnableAdminServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-@EnableAdminServer
-
-public class EurekaServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaServerApplication.class, args);
-	}
-}
