+++ eureka-server/src/main/java/com/service/eureka/EurekaServerApplication.java
@@ -1,0 +1,21 @@
+package com.service.eureka;
+
+import de.codecentric.boot.admin.config.EnableAdminServer;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+
+
+@SpringBootApplication
+@EnableEurekaServer
+@EnableAdminServer
+@Slf4j
+public class EurekaServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EurekaServerApplication.class, args);
+        log.debug("Eureka Server Application started..!!");
+    }
+}
